auth to all mutations

+++ src/resolvers/Post/DeletePost.ts
@@ -5,9 +5,13 @@ import {
   PubSub,
   PubSubEngine,
   Resolver,
+  UseMiddleware,
 } from 'type-graphql'
-import { MutationType } from '../../enums/mutationType'
-import { Topics } from '../../enums/subscriptions'
+import { models } from '../../types/enums/models'
+import { MutationType } from '../../types/enums/mutationType'
+import { Topics } from '../../types/enums/subscriptions'
+import { isAuth } from '../../middleware/isAuth'
+import { IsOwner } from '../shared/auth/isOwner'
 import { Post } from '../../models/Post'
 import { MyContext } from '../../types/MyContext'
 import { Select } from '../shared/selectParamDecorator'
@@ -20,12 +24,25 @@ const { Posts } = Topics
 @Resolver()
 class DeletePostResolver {
   @Mutation(() => Post)
+  @UseMiddleware(isAuth)
+  @IsOwner(models.post)
   async deletePost(
     @Args() { id }: PostIdInput,
-    @Ctx() { prisma }: MyContext,
+    @Ctx() { prisma, userId }: MyContext,
     @PubSub() pubSub: PubSubEngine,
     @Select() select: any
   ): Promise<Post> {
+    const isOwner = !!(
+      await prisma.post.findMany({
+        where: { id, authorId: userId },
+        select: { id: true },
+      })
+    )[0]
+
+    if (!isOwner) {
+      throw new Error("You're Not the owner of the Post!")
+    }
+
     const deletedPost = (await prisma.post.delete({
       where: { id },
       select: { ...select, id: true },
